adding first draft of ICON image analysis
